+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -24,6 +24,8 @@
 * Added support for setting the username and password in basic auth pop-up
    dialogs (currently IE only).
 * Deprecated `WebElement#getInnerHtml()` and `WebEleemnt#getOuterHtml()`
+* Deprecated `Promise#thenCatch()` - use `Promise#catch()` instead
+* Deprecated `Promise#thenFinally()` - use `promise.thenFinally()` instead
 * FIXED: `io.findInPath()` will no longer match against directories that have
    the same basename as the target file.
 * FIXED: `phantomjs.Driver` now takes a third argument that defines the path to